Completed task 11 Hussain & Hesham

+++ restaurants/views.py
@@ -2,6 +2,7 @@ from django.shortcuts import render, redirect
 from .models import Restaurant, Item
 from .forms import RestaurantForm, ItemForm, SignupForm, SigninForm
 from django.contrib.auth import login, authenticate, logout
+from django.http import HttpResponse
 
 def signup(request):
     form = SignupForm()
@@ -42,6 +43,10 @@ def signout(request):
     logout(request)
     return redirect("signin")
 
+def no_access(request):
+    return HttpResponse("YOU HAVE NO ACCESS!!!!!")
+# return render(request, 'no_access.html')
+
 def restaurant_list(request):
     context = {
         "restaurants":Restaurant.objects.all()
@@ -59,6 +64,8 @@ def restaurant_detail(request, restaurant_id):
     return render(request, 'detail.html', context)
 
 def restaurant_create(request):
+    if request.user.is_anonymous:
+        return redirect('signin')
     form = RestaurantForm()
     if request.method == "POST":
         form = RestaurantForm(request.POST, request.FILES)
@@ -75,13 +82,16 @@ def restaurant_create(request):
 def item_create(request, restaurant_id):
     form = ItemForm()
     restaurant = Restaurant.objects.get(id=restaurant_id)
-    if request.method == "POST":
-        form = ItemForm(request.POST)
-        if form.is_valid():
-            item = form.save(commit=False)
-            item.restaurant = restaurant
-            item.save()
-            return redirect('restaurant-detail', restaurant_id)
+    if request.user.is_staff or request.user == restaurant.owner:
+        if request.method == "POST":
+            form = ItemForm(request.POST)
+            if form.is_valid():
+                item = form.save(commit=False)
+                item.restaurant = restaurant
+                item.save()
+                return redirect('restaurant-detail', restaurant_id)
+    else:
+        return redirect('no-access')
     context = {
         "form":form,
         "restaurant": restaurant,
@@ -90,12 +100,15 @@ def item_create(request, restaurant_id):
 
 def restaurant_update(request, restaurant_id):
     restaurant_obj = Restaurant.objects.get(id=restaurant_id)
-    form = RestaurantForm(instance=restaurant_obj)
-    if request.method == "POST":
-        form = RestaurantForm(request.POST, request.FILES, instance=restaurant_obj)
-        if form.is_valid():
-            form.save()
-            return redirect('restaurant-list')
+    if request.user.is_staff or request.user == restaurant_obj.owner:
+        form = RestaurantForm(instance=restaurant_obj)
+        if request.method == "POST":
+            form = RestaurantForm(request.POST, request.FILES, instance=restaurant_obj)
+            if form.is_valid():
+                form.save()
+                return redirect('restaurant-list')
+    else:
+        return redirect('no-access')
     context = {
         "restaurant_obj": restaurant_obj,
         "form":form,
@@ -104,5 +117,22 @@ def restaurant_update(request, restaurant_id):
 
 def restaurant_delete(request, restaurant_id):
     restaurant_obj = Restaurant.objects.get(id=restaurant_id)
-    restaurant_obj.delete()
-    return redirect('restaurant-list')
+    if request.user.is_staff:
+        restaurant_obj.delete()
+        return redirect('restaurant-list')
+    return redirect('no-access')
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
